Baseline commit for PWM motor controller
Need to finish the hardware build to test but this gets our ready to test code saved.  In software engineering terms, this is "code complete"; we've completed all of the coding and no features or functionality remains to be developed.

Code changes will3likely still occur as a result of testing which will happen next.