Corrected timing values for PWM outputs
Correct servo "angle" representations to stay with 1.0-2.0ms pulse width range as well as getting as close to 1.5ms "stopped" state as possible (achieved a "stop" pulse width of 1.504ms, full reverse at 1.018ms, and full forward at 1.998ms).

Improved some of the documentation and improved some of the debugging output (always show the potentiometer value even if the switch is in the "stop" state).